Add allure steps to header/footer
