+++ component/client/HeaderChangeLangBtn.ts
@@ -1,13 +1,12 @@
 import {Locator, Page} from "@playwright/test";
 import {BaseComponent} from "./BaseComponent";
 import {LanguageDetails} from "../../data/enums";
-
+import * as allure from "allure-js-commons";
 
 export class HeaderChangeLangBtn extends BaseComponent {
     image: Locator;
     menuLocator: Locator;
     checkSelected: Locator;
-
 
     constructor(page: Page, parent: Locator) {
         super(page, parent);
@@ -17,24 +16,37 @@ export class HeaderChangeLangBtn extends BaseComponent {
     }
 
     async click(): Promise<void> {
-        await this.image.click();
+        await allure.step('Click Change Language Btn', async () => {
+            await this.image.click();
+        })
     }
 
     async getImage(): Promise<string> {
-        return await this.image.getAttribute('src') || "";
+        let link: string = "";
+            await allure.step('Get image link of Change Language Btn', async () => {
+            link = await this.image.getAttribute('src') || "";
+        })
+        return link;
     }
 
     async waitForHeaderChangeLangBtnVisible(): Promise<void> {
-        await this.waitIsVisible(this.menuLocator);
+        await allure.step(`Check if Change Language Btn is visible`, async () => {
+            await this.waitIsVisible(this.menuLocator);
+        })
     }
 
     async selectLanguage(language: LanguageDetails): Promise<void> {
-
-        await this.menuLocator.locator(`li:has-text("${language.value}")`).click();
-        await this.page.waitForURL(`**/${language.shortName}/**`, { waitUntil: 'domcontentloaded' });
+        await allure.step(`Select Language ${language}`, async () => {
+            await this.menuLocator.locator(`li:has-text("${language.value}")`).click();
+            await this.page.waitForURL(`**/${language.shortName}/**`, { waitUntil: 'domcontentloaded' });
+        })
     }
 
     async getSelectedLanguage(): Promise<string> {
-        return (await this.checkSelected.innerText());
+        let language: string = '';
+        await allure.step(`Get selected Language`, async () => {
+            language = await this.checkSelected.innerText();
+        })
+        return language;
     }
 }
