+++ config/env.ts
@@ -4,7 +4,7 @@ import path from 'path';
 // Load .env from repository root by default
 dotenv.config({ path: path.resolve(__dirname, '..', '.env') });
 
-export const BASE_CLIENT_URL = process.env.BASE_CLIENT_URL ?? 'https://lf-client-stage-a3ama9eydjfucnbj.polandcentral-01.azurewebsites.net/';
+export const BASE_CLIENT_URL = process.env.BASE_CLIENT_URL ?? 'https://lf-client-qa-stage-atbchmhfgtaxfdas.polandcentral-01.azurewebsites.net/';
 export const BASE_ADMIN_URL = process.env.BASE_ADMIN_URL ?? `https://lf-admin-qa-stage-apbug0eka2bsdkf9.polandcentral-01.azurewebsites.net/`;
 export const BASE_API_URL = process.env.BASE_API_URL ?? 'https://jsonplaceholder.typicode.com';
