+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "test": "playwright test",
     "test:ui": "playwright test tests/ui.spec.ts",
+    "client:ui":"playwright test tests/client",
     "test:api": "playwright test tests/api.spec.ts",
     "typecheck": "tsc -p tsconfig.json --noEmit",
     "show": "playwright show-report",
