+++ page/client/BasePage.ts
@@ -2,7 +2,7 @@ import {Locator, Page} from '@playwright/test';
 import { HeaderComponent } from '../../component/client/HeaderComponent';
 import { FooterComponent } from '../../component/client/FooterComponent';
 import {CookiesModal} from "../../component/client/CookiesModal";
-import {allure} from "allure-playwright";
+import * as allure from "allure-js-commons";
 
 export class BasePage {
   protected page: Page;
@@ -21,12 +21,12 @@ export class BasePage {
 
   async goto(path: string): Promise<void> {
       await this.page.goto(path, { waitUntil: 'domcontentloaded' });
-      await this.cookiesModal.acceptAll(2000);
+      await this.cookiesModal.acceptAll(10000);
   }
 
   async getTitleText(): Promise<string> {
     let text = '';
-    await allure.step('Navigate to Playwright website', async () => {
+    await allure.step('Get Page title', async () => {
       text =  await this.title.textContent() || '';
     });
     return text;
